Seperate cli.py into multiple files
