+++ aishell/cli/__init__.py
@@ -1,0 +1,3 @@
+import typer
+
+cli_app = typer.Typer()
